Consolidate types into a single file and import as needed
